nsswitch: Add group merging support
https://sourceware.org/glibc/wiki/Proposals/GroupMerging

== Justification ==
It is common today for users to rely on centrally-managed user stores for
handling their user accounts. However, much software existing today does not
have an innate understanding of such accounts. Instead, they commonly rely on
membership in known groups for managing access-control (for example the "wheel"
group on Fedora and RHEL systems or the "adm" group on Debian-derived systems).
In the present incarnation of nsswitch, the only way to have such groups
managed by a remote user store such as FreeIPA or Active Directory would be to
manually remove the groups from /etc/group on the clients so that nsswitch
would then move past nss_files and into the SSSD, nss-ldap or other remote user
database.

== Solution ==
With this patch, a new action is introduced for nsswitch: NSS_ACTION_MERGE.
To take advantage of it, one will add [SUCCESS=merge] between two database
entries in the nsswitch.conf file. When a group is located in the first of the
two group entries, processing will continue on to the next one. If the group is
also found in the next entry (and the group name and GID are an exact match),
the member list of the second entry will be added to the group object to be
returned.

== Implementation ==
After each DL_LOOKUP_FN() returns, the next action is checked. If the function
returned NSS_STATUS_SUCCESS and the next action is NSS_ACTION_MERGE, a copy of
the result buffer is saved for the next pass through the loop. If on this next
pass through the loop the database returns another instance of a group matching
both the group name and GID, the member list is added to the previous list and
it is returned as a single object. If the following database does not contain
the same group, then the original is copied back into the destination buffer.

This patch implements merge functionality only for the group database. For
other databases, there is a default implementation that will return the EINVAL
errno if a merge is requested. The merge functionality can be implemented for
other databases at a later time if such is needed. Each database must provide a
unique implementation of the deep-copy and merge functions.

== No Premature Optimizations ==
The following is a list of places that might be eligible for optimization, but
were not overengineered for this initial contribution:
 * Any situation where a merge may occur will result in one malloc() of the
   same size as the input buffer.
 * Any situation where a merge does occur will result in a second malloc() to
   hold the list of pointers to member name strings.
 * The list of members is simply concatenated together and is not tested for
   uniqueness (which is identical to the behavior for nss_files, which will
   simply return identical values if they both exist on the line in the file.
   This could potentially be optimized to reduce space usage in the buffer, but
   it is both complex and computationally expensive to do so.

== Known Bugs ==
The merging function does not work with enumerations (such as getgrent()). This
is because those routines are all sequential and we cannot go back and change
the output they have already produced. The only way to implement group merging
on enumeration would be for glibc to process all possible results ahead of time
and cache them. This is far beyond the scope of this patch.

Mitigation: Most applications in the real world do not rely on getgrent() to
retrieve memberships. They may use it to list the available groups by name and
ID, but most will interrogate the group entries individually to get the full
list of members. Additionally, many centrally-managed identity sources
(including the SSSD and Winbind) disable user and group enumeration by default
due to its inherently slow behavior when operating over the network. As a
result, client software already needs to be tolerant of incomplete enumeration
data.

Resolves BZ #19072